+++ Mods ChoGGi/Forestation Plant Goes To 11/metadata.lua
@@ -1,0 +1,18 @@
+return PlaceObj("ModDef", {
+	"title", "Forestation Plant Goes To 11",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"saved", 1558267200,
+	"image", "Preview.png",
+	"id", "ChoGGi_ForestationPlantGoesTo11",
+	"steam_id", "1745954351",
+	"pops_any_uuid", "a61d60d3-ea45-4d67-94f3-0e3f866cb967",
+	"author", "ChoGGi",
+	"lua_revision", 244677,
+	"code", {
+		"Code/Script.lua",
+	},
+	"description", [[The plant doesn't stop at 40% anymore.]],
+	"TagTerraforming", true,
+})