+++ Expanded Cheat Menu/Changelog.md
@@ -3,6 +3,8 @@
 ## v13.6 (Unreleased)
 ### Fixed:
 - Anomaly Scanning didn't open (thanks mys721tx).
+- Consts and Service Building Stats (thanks Kami-sama).
+- Set Gender = Other made colonists invisible (thanks Kami-sama).
 
 ## v13.5 (May 17 2019)
 ### Fixed: